pyproject.toml
# RAG Service MVP - AI Component Validation Platform
# Web service for validating AI development components with RAG capabilities

[project]
name = "rag-service"
version = "0.1.0"
description = "RAG Service MVP for AI Component Validation Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "OA Component Team" }
]
keywords = ["rag", "ai", "llm", "milvus", "langfuse", "phidata", "litellm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # AI/LLM
    "phidata>=0.10.0",
    "litellm>=1.40.0",
    "openai>=1.14.0",
    # Observability
    "langfuse>=2.32.0",
    # Vector Database
    "pymilvus>=2.3.6",
    # Embeddings
    "sentence-transformers>=2.7.0",
    # HTTP Client
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.3.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/oa-component/rag-service"
Repository = "https://github.com/oa-component/rag-service"
Issues = "https://github.com/oa-component/rag-service/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_service"]

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "slow: Slow-running tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if 0:",
    "if False:",
]
precision = 2

[tool.coverage.paths]
source = [
    "src/",
]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",  # Unused arguments are common in test fixtures
    "S101", # assert is fine in tests
]

[tool.ruff.lint.isort]
known-first-party = ["rag_service"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# MyPy Configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "phidata.*",
    "litellm.*",
    "langfuse.*",
    "pymilvus.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true
